Se separa el regisrto Tramite
