+++ src/app/SCT/pages/registro-tamite/registro-tamite.component.html
@@ -2,102 +2,6 @@
     <p-toast position="bottom-center"></p-toast>
     <p-card class="col-9 mx-auto">
         <h1 class="fbs-roboto-300 mb-4">Registro Tramite</h1>
-        <form  class="p-fluid" [formGroup]="registroTramiteForm" autocomplete="off" >
-
-            <div class="row mb-3">
-                <div class="col-lg-6 col-md-12">
-                    <label class="fbs-roboto-300">Cédula cliente</label>
-                    <div class="p-inputgroup">
-                        <input type="text"
-                            #inputCedula
-                            pInputText
-                            pAutoFocus
-                            [autofocus]="true"
-                            formControlName="CEDULA"
-                            (keyup.enter)="buscar()">
-                            
-                        <span class="p-inputgroup-addon"><i class="pi pi-search" (click)="buscar()"></i></span>
-                    </div>
-                    <small class="p-error block" *ngIf="esInvalido('CEDULA')">
-                        <i class="pi pi-exclamation-circle"></i>
-                        Ingrese una cédula válida
-                    </small>
-                </div>
-                <div class="col-lg-6 col-md-12">
-                    <label class="fbs-roboto-300" >Tipo de Afiliación</label>
-                    <span class="p-float-label">
-                        <input class="form-control shadow disabled" type="text" pInputText 
-                            formControlName="TIPO_CLIENTE">
-                    </span>
-                </div>
-            </div>
-        
-            <div class="row mb-3">
-                <div class="col-lg-4 col-md-6 col-sm-12">
-                    <label class="fbs-roboto-300" for="float-input">Nombre</label>
-                    <span class="p-float-label">
-                        <input id="float-input" type="text" pInputText
-                        formControlName="NOMBRE"
-                        class="form-control shadow disabled">
-                    </span>
-                </div>
-                <div class="col-lg-4 col-md-6 col-sm-12">
-                    <label class="fbs-roboto-300" for="float-input">Apellido 1</label>
-                    <span class="p-float-label">
-                        <input id="float-input" type="text" pInputText
-                        formControlName="APELLIDO_1"
-                        class="form-control shadow disabled">
-                    </span>
-                </div>
-                <div class="col-lg-4 col-md-6 col-sm-12">
-                    <label class="fbs-roboto-300" for="float-input">Apellido 2</label>
-                    <span class="p-float-label">
-                        <input id="float-input" type="text" pInputText
-                        formControlName="APELLIDO_2"
-                        class="form-control shadow disabled">
-                    </span>
-                </div>
-            </div>
-
-            <div class="row mb-3">
-                <div class="col-lg-12 col-md-12">
-                    <label class="fbs-roboto-300" for="float-input">Trámite Realizado</label>
-                    <div>
-                        <p-dropdown 
-                            [options]="nombreTramites"
-                            [editable]="true"
-                            optionLabel="NOMBRE_TRAMITE" 
-                            optionValue="ID_TRAMITE"
-                            placeholder="Seleccione Trámite"
-                            formControlName="TRAMITE">
-                        </p-dropdown>
-                    </div>
-                    <small class="p-error block" *ngIf="esInvalido('TRAMITE')">
-                        <i class="pi pi-exclamation-circle"></i>
-                        Favor seleccionar el tramite realizado
-                    </small>
-                </div>
-            </div>
-    
-            <div class="row mb-2 mx-auto">
-                <label class="fbs-roboto-300">Descripción</label>
-                <textarea rows="3" cols="30" pInputTextarea [autoResize]="true" formControlName="DESCRIPCION"></textarea>
-            </div>
-        
-            <div class="row mb-1">
-                <div class="col">
-                    <div class="row float-end">
-                        <div class="col">
-                            <p-button label="Limpiar" styleClass="p-button-danger" icon="pi pi-trash"></p-button>
-                        </div>
-                        <div class="col">
-                            <p-button label="Registrar" styleClass="p-button-warning" icon="pi pi-check" (click)="agregarRegistroTramite()"></p-button>
-                            
-                        </div>
-                    </div>
-                </div>
-            </div>
-        
-        </form>
+        <app-formulario-registro-tramite></app-formulario-registro-tramite>
     </p-card>
 </div>
